# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SwingQuant TITAN API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/universe/rebuild": {
      "post": {
        "tags": [
          "universe"
        ],
        "summary": "Rebuild Quantamental Universe",
        "description": "Rafraîchit l'univers quantamental en subprocess détaché.\n\nDeux modes (mode=\"staggered\" par défaut) :\n  • staggered → universe_scheduler : budget 50 tickers/jour, skip ceux\n    refreshés < 14j, merge non-destructif. **Quota-friendly** (~150 calls\n    FMP sur 250/j disponibles). Mode recommandé pour les usages réguliers.\n  • full → universe_engine : rebuild intégral (~500 tickers × 3 calls =\n    1500 calls FMP). Crame le quota journalier → **à réserver aux\n    urgences** (corruption universe.json, nouveau déploiement).\n\nTéléchargement lourd → renvoie immédiatement {job_id} pour polling\nvia GET /api/jobs/{job_id}. Refuse (409) si un job est déjà en cours.",
        "operationId": "rebuild_quantamental_universe_api_universe_rebuild_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniverseRebuildRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobLaunchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/universe": {
      "get": {
        "tags": [
          "universe"
        ],
        "summary": "Get Quantamental Universe",
        "description": "SOURCE OF TRUTH — univers quantamental long-terme.\nLit data/universe.json. Payload : tickers enrichis (sector, market_cap,\nforward_pe, price targets, recommandations analystes), agrégation\nsectorielle, stats du dernier rebuild.\n\nFiltre optionnel ?sector=Technology pour restreindre la réponse.\n\nConditional GET : émet un ETag basé sur la mtime de universe.json. Le\nclient peut renvoyer `If-None-Match` → 304 Not Modified (zéro payload)\ntant que le rebuild n'a pas tourné. Le filtre `sector` est inclus dans\nl'ETag car il change la réponse.\n\nMétadonnées W8 :\n  - last_updated_timestamp / stale_days / is_stale : âge du cache\n  - is_incomplete par ticker (flag injecté dans chaque entrée)\n  - incomplete_count / incomplete_ratio : agrégat global",
        "operationId": "get_quantamental_universe_api_universe_get",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniverseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sectors": {
      "get": {
        "tags": [
          "sectors"
        ],
        "summary": "Get Sectors",
        "description": "Liste des 11 secteurs GICS enrichis :\n  • Agrégats robustes (médianes/std) depuis universe.json\n  • Distribution des recommandations analystes\n  • Top-3 tickers par upside et market cap\n  • Momentum 6M via ETF sectoriel SPDR (cache 24h)\n  • Rotation Score = 0.4·valuation + 0.3·sentiment + 0.3·upside\n    (composantes normalisées min-max cross-sector, scaled 0-100)\n\nQuery params :\n  refresh_momentum=true  → force re-download yfinance (bypass cache).\n                           Requiert Authorization: Bearer (anti-ban IP).",
        "operationId": "get_sectors_api_sectors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "refresh_momentum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh Momentum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sectors/{sector}": {
      "get": {
        "tags": [
          "sectors"
        ],
        "summary": "Get Sector Detail",
        "description": "Détail d'un secteur : stats agrégées + tous ses tickers enrichis\n(upside_pct, reco_bucket) pour un rendu direct dans la vue drill-down.",
        "operationId": "get_sector_detail_api_sectors__sector__get",
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobLaunchResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "job": {
            "$ref": "#/components/schemas/JobMeta"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "ok",
          "job"
        ],
        "title": "JobLaunchResponse"
      },
      "JobMeta": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cmd": {
            "type": "string",
            "title": "Cmd"
          },
          "pid": {
            "type": "integer",
            "title": "Pid"
          },
          "started": {
            "type": "string",
            "title": "Started"
          },
          "log_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Path"
          },
          "running": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Running"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "job_id",
          "name",
          "cmd",
          "pid",
          "started"
        ],
        "title": "JobMeta"
      },
...
      "UniverseRebuildRequest": {
        "properties": {
          "indices": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indices"
          },
          "min_market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Market Cap"
          },
          "workers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workers"
          },
          "max_retries": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Retries"
          },
          "timeout": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "min_age_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Age Days"
          },
          "force": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Force"
          }
        },
        "type": "object",
        "title": "UniverseRebuildRequest",
        "description": "Payload optionnel pour POST /api/universe/rebuild."
      },
      "UniverseResponse": {
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "last_updated_timestamp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated Timestamp"
          },
          "stale_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Days"
          },
          "is_stale": {
            "type": "boolean",
            "title": "Is Stale",
            "default": false
          },
          "staleness_threshold_days": {
            "type": "integer",
            "title": "Staleness Threshold Days",
            "default": 7
          },
          "incomplete_count": {
            "type": "integer",
            "title": "Incomplete Count",
            "default": 0
          },
          "incomplete_ratio": {
            "type": "number",
            "title": "Incomplete Ratio",
            "default": 0.0
          },
          "source_indices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Indices"
          },
          "filter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filter"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          },
          "sectors": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sectors"
          },
          "tickers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Tickers"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "rebuilding": {
            "type": "boolean",
            "title": "Rebuilding"
          },
          "rebuild_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rebuild Job Id"
          },
          "last_change": {
            "title": "Last Change"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "count",
          "rebuilding"
        ],
        "title": "UniverseResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}